requirements.txt
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.9
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flower==2.0.1
hiredis==2.3.2
humanize==4.9.0
inflection==0.5.1
kombu==5.3.4
packaging==23.2
pillow==10.2.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tornado==6.4
tzdata==2023.4
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.12
